We have working code!!! woohoo party!! Check out the sandbox to run some code and see how it's going!! I also deleted some unused/duplicate files and cleaned some stuff up.

+++ Items/OneTimeExpense.java
@@ -1,8 +1,7 @@
 class OneTimeExpense extends Item{
 
-    public OneTimeExpense(String n, int v){
+    public OneTimeExpense(String n, float v){
         super(n, v);
     }
-
 
 }
