+++ Items/OneTimeIncome.java
@@ -1,6 +1,6 @@
 class OneTimeIncome extends Item{
 
-    public OneTimeIncome(String n, int v){
+    public OneTimeIncome(String n, float v){
         super(n, v);
     }
