+++ Items/RecurringExpense.java
@@ -1,6 +1,6 @@
 class RecurringExpense extends Item{
 
-    public RecurringExpense(String n, int v, int f){
+    public RecurringExpense(String n, float v, int f){
         super(n, v, f);
     }
