+++ Items/RecurringIncome.java
@@ -1,6 +1,6 @@
 class RecurringIncome extends Item{
     
-    public RecurringIncome(String n, int v, int f){
+    public RecurringIncome(String n, float v, int f){
         super(n, v, f);
     }
